+++ aws-opentelemetry-distro/tests/amazon/opentelemetry/distro/patches/test_instrumentation_patch.py
@@ -25,7 +25,6 @@ from amazon.opentelemetry.application_signals.internal.semconv.aws_attributes im
     AWS_REGION,
     AWS_SQS_QUEUE_NAME,
 )
-from amazon.opentelemetry.distro.patches._instrumentation_patch import apply_instrumentation_patches
 from amazon.opentelemetry.application_signals.internal.semconv.gen_ai_attributes import (
     GEN_AI_BROWSER_ID,
     GEN_AI_CODE_INTERPRETER_ID,
@@ -33,6 +32,7 @@ from amazon.opentelemetry.application_signals.internal.semconv.gen_ai_attributes
     GEN_AI_MEMORY_ID,
     GEN_AI_RUNTIME_ID,
 )
+from amazon.opentelemetry.distro.patches._instrumentation_patch import apply_instrumentation_patches
 from opentelemetry.instrumentation.botocore import BotocoreInstrumentor
 from opentelemetry.instrumentation.botocore.extensions import _BOTOCORE_EXTENSIONS, bedrock_utils
 from opentelemetry.propagate import get_global_textmap