+++ scripts/build_and_install_distro.sh
@@ -16,9 +16,9 @@ fi
 # Setup - update dependencies and create/empty dist dir
 python3 -m pip install --upgrade pip setuptools wheel packaging build
 mkdir -p dist
-rm -rf dist/aws_opentelemetry_distro* dist/aws_opentelemetry_application_signals* dist/aws_opentelemetry_serviceevents*
+rm -rf dist/aws_opentelemetry_distro* dist/aws_opentelemetry_components*
 
-for pkg in aws-opentelemetry-application-signals aws-opentelemetry-serviceevents aws-opentelemetry-distro; do
+for pkg in aws-opentelemetry-components aws-opentelemetry-distro; do
   (cd "$pkg" && python3 -m build --outdir ../dist)
 done